+++ jest.coverage.ignore.js
@@ -6,17 +6,18 @@ const PERMANENT_IGNORE = [
   './pages/_app.tsx',
   './ui/index.ts',
   './packages/styleguide/src/StyleGuide.tsx',
+  './projects/gtms-frontend/packages/ui/Forms/index.ts',
 ]
 
 const TEMPORARY_IGNORE = [
-  './components/common/FlipCard/index.tsx',
-  './pages/registration/index.tsx',
-  './hooks/auth.ts',
-  './pages/index.tsx',
-  './api/auth/googleLogin.ts',
-  './components/login/SocialButtons/index.tsx',
   './packages/ui/PostSingle/index.tsx',
   './packages/app-andrew/src/pages/group/index.tsx',
+  './packages/app-andrew/src/server.ts',
+  './packages/ui/SavedSearch/index.tsx',
+  './packages/app-main/src/pages/rules/index.tsx',
+  './packages/app-main/src/providers/Rules/index.tsx',
+  './packages/app-main/src/components/rules/AcceptRulesButton/index.tsx',
+  './projects/gtms-frontend/packages/ui/Forms/ExpandingTextarea/index.tsx',
 ]
 
 module.exports = [...PERMANENT_IGNORE, ...TEMPORARY_IGNORE]