+++ packages/app-andrew/src/components/login/SocialButtons/index.tsx
@@ -15,7 +15,7 @@ export const SocialButtons: FC<{
     appId: process.env.FB_APP_ID,
     fields: 'name, email, picture',
     onFailure,
-    onSuccess: async payload => {
+    onSuccess: async (payload) => {
       const { accessToken, id } = payload
 
       setIsLoading(true)
@@ -53,7 +53,7 @@ export const SocialButtons: FC<{
             <GoogleLogin
               clientId={process.env.GOOGLE_CLIENT_ID}
               buttonText="Google"
-              onSuccess={async response => {
+              onSuccess={async (response) => {
                 setIsLoading(true)
 
                 const {