+++ packages/app-andrew/src/pages/_app.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import App from 'next/app'
-import RulesProvider from '../providers/Rules'
 import '@gtms/styles/scss/global.scss'
 import { appWithTranslation } from '@gtms/commons/i18n'
 import { NavigationDot } from '@gtms/ui/NavigationDot'
@@ -10,10 +9,10 @@ class GTMSApp extends App {
     const { Component, pageProps } = this.props
 
     return (
-      <RulesProvider>
+      <div>
         <NavigationDot />
         <Component {...pageProps} />
-      </RulesProvider>
+      </div>
     )
   }
 }