+++ packages/app-main/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.3.0 (2020-03-28)
+
+
+### Bug Fixes
+
+* fixes for ts errors ([#220](https://github.com/gtms-org/gtms-frontend/issues/220)) ([3467724](https://github.com/gtms-org/gtms-frontend/commit/34677245f46984cc12c11f85e61795510a15c740))
+* GEOT-79 remove unneeded asyncs ([#225](https://github.com/gtms-org/gtms-frontend/issues/225)) ([3b68471](https://github.com/gtms-org/gtms-frontend/commit/3b6847171c5a30515dc42c4343793ffac67ce6a7))
+
+
+### Features
+
+* GEOT-42: group main page example ([#312](https://github.com/gtms-org/gtms-frontend/issues/312)) ([c7dd7cd](https://github.com/gtms-org/gtms-frontend/commit/c7dd7cdcc45a2b51403ce066c8fdce11b54a2e99))
+* GEOT-64: group create ([#323](https://github.com/gtms-org/gtms-frontend/issues/323)) ([1516233](https://github.com/gtms-org/gtms-frontend/commit/1516233651b28f40e36145ae7cacb37867e6ef45))
+* GEOT-73: GEOT-73: font-weight global solution ([#333](https://github.com/gtms-org/gtms-frontend/issues/333)) ([4c6bab2](https://github.com/gtms-org/gtms-frontend/commit/4c6bab2b8713a1fad0fc219d48128feda79dac32))
+* monorepo migration ([#212](https://github.com/gtms-org/gtms-frontend/issues/212)) ([3400d84](https://github.com/gtms-org/gtms-frontend/commit/3400d84f411612076adba5920af8b323b55f473a))
+
+
+
+
+
+## [0.2.1](https://github.com/gtms-org/gtms-frontend/compare/@gtms/app-main@0.2.0...@gtms/app-main@0.2.1) (2020-03-23)
+
+**Note:** Version bump only for package @gtms/app-main
+
+
+
+
+
 # [0.2.0](https://github.com/gtms-org/gtms-frontend/compare/@gtms/app-main@0.1.5...@gtms/app-main@0.2.0) (2020-03-23)