+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.2.0 (2020-03-28)
+
+
+### Features
+
+* GEOT-107: load global css into styleguide ([#263](https://github.com/gtms-org/gtms-frontend/issues/263)) ([b30da1b](https://github.com/gtms-org/gtms-frontend/commit/b30da1b131397f03159e95f67ccd74c3786bea3c))
+
+
+
+
+
 ## [0.1.3](https://github.com/gtms-org/gtms-frontend/compare/@gtms/styleguide@0.1.2...@gtms/styleguide@0.1.3) (2020-03-22)
 
 **Note:** Version bump only for package @gtms/styleguide