+++ packages/styles/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.3.0 (2020-03-28)
+
+
+### Features
+
+* GEOT-116 user name change form ([#271](https://github.com/gtms-org/gtms-frontend/issues/271)) ([09dac18](https://github.com/gtms-org/gtms-frontend/commit/09dac183e0bbe1137e25f23f651b731d9ea94bef))
+* GEOT-42: group main page example ([#312](https://github.com/gtms-org/gtms-frontend/issues/312)) ([c7dd7cd](https://github.com/gtms-org/gtms-frontend/commit/c7dd7cdcc45a2b51403ce066c8fdce11b54a2e99))
+* GEOT-64: group create ([#323](https://github.com/gtms-org/gtms-frontend/issues/323)) ([1516233](https://github.com/gtms-org/gtms-frontend/commit/1516233651b28f40e36145ae7cacb37867e6ef45))
+* GEOT-73: GEOT-73: font-weight global solution ([#333](https://github.com/gtms-org/gtms-frontend/issues/333)) ([4c6bab2](https://github.com/gtms-org/gtms-frontend/commit/4c6bab2b8713a1fad0fc219d48128feda79dac32))
+* GEOT-74 user account page ([#223](https://github.com/gtms-org/gtms-frontend/issues/223)) ([1b9c005](https://github.com/gtms-org/gtms-frontend/commit/1b9c00586ea8990576fc059e0cd7707811194001))
+* monorepo migration ([#212](https://github.com/gtms-org/gtms-frontend/issues/212)) ([3400d84](https://github.com/gtms-org/gtms-frontend/commit/3400d84f411612076adba5920af8b323b55f473a))
+
+
+
+
+
 # [0.2.0](https://github.com/gtms-org/gtms-frontend/compare/@gtms/styles@0.1.2...@gtms/styles@0.2.0) (2020-03-23)